Fix leakage of temporary directory
As it turns out, the temporary directory that we create for our FIFO
objects gets leaked, because of our usage of std::process::exit(), which
does not invoke object destructors properly.
Fix the issue by exiting from a different function scope, guaranteeing
that all objects of the "inner" function have been destroyed.